+++ packages/@groupby/elements-cache-driver-plugin/src/index.ts
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
-export { default as CacheDriverPlugin } from './cache-driver-plugin';
+export { default as CacheDriverPlugin, CacheDriverOptions } from './cache-driver-plugin';