+++ packages/@groupby/elements-sayt-driver-plugin/src/index.ts
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/prefer-default-export
-export { default as SaytDriverPlugin } from './sayt-driver-plugin';
+export { default as SaytDriverPlugin, SaytDriverOptions } from './sayt-driver-plugin';